Rename hooks/useDataDog.ts to lib/dataDog.ts
useDataDog did not export any hooks.
